requirements.txt
Django
jsonfield
PyGithub
djangorestframework
django-filter